server:#30:refactoring and debugging v1

+++ controllers/post/postUnlike.js
@@ -11,47 +11,36 @@ module.exports = async (req, res) => {
   };
   const { post_id } = req.params;
 
-  await post.findOne({
+  const valid_post = await post.findOne({
     where: {
       id: post_id
     }
   })
-  .then((result) => {
-    if(!result) {
-      return res.status(404).json({
-        data: null,
-        message: '내용을 찾을 수 없습니다.'
-      })
-    }
-  })
   .catch((err) => {
-    res.status(500).json({
-      data: null,
-      message: 'Server Error'
-    })
+    console.error(err)
   });
+  if(!valid_post) {
+    return res.status(404).json({
+      data: null,
+      message: '내용을 찾을 수 없습니다.'
+    })
+  }
 
   const user_id = await user.findOne({
     where: {
       username: userData.username
     }
   })
-  .then((result) => {
-    if(!result) {
-      return res.status(403).json({
-        data: null,
-        message: 'You need sign up'
-      })
-    }
-    return result.dataValues.id;
-  })
   .catch((err) => {
-    res.status(500).json({
-      data: null,
-      message: 'Server Error'
-    })
+    console.error(err)
   });
-
+  if(!user_id){
+    return res.status(403).json({
+      data: null,
+      message: 'You need sign up'
+    })
+  }
+  user_id = user_id.dataValues.id
   await post_like.destroy({
     where: {
       user_id,
@@ -71,9 +60,6 @@ module.exports = async (req, res) => {
     });
   })
   .catch((err) => {
-    res.status(500).json({
-      data: null,
-      message: 'Server Error'
-    })
+    console.error(err)
   });
 }
